pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling >= 1.26"]

[project]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  "py-opensonic == 7.0.2",
  "tqdm == 4.67.1",
  "python-magic == 0.4.27",
]
description = "A small example package"
license = "MIT"
license-files = ["license.md"]
name = "downsonic"
readme = "readme.md"
requires-python = ">=3.13.9"
version = "0.0.1"

[[project.authors]]
email = "person@dandelion.computer"
name = "andrew sòlanto"

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
downsonic = "downsonic.main:main"
